Implement the silver searcher to be used by ctrlp.vim
From https://thoughtbot.com/upcase/videos/onramp-to-vim-plugins:
By default CtrlP will cache the file listing it uses to display results.
This improves the speed of matching but means that the file listing can
become stale and show incorrect results.

Instead, we recommend using the following configuration to instruct
CtrlP to use ag, aka the silver searcher, to populate the file list and
to never cache the results. This is both amazingly fast, will always be
up to date since is does not cache, and will respect your .gitignore
file to ignore build and tmp files.

=============================================

https://github.com/ggreer/the_silver_searcher
https://github.com/r00k/dotfiles/blob/master/vimrc