Move all dotfiles into specific folder
This repository had been initialized as a bare repo.
Main advantages consisted in removing needs for extra
tooling or symlinks creation.
Reference:
https://developer.atlassian.com/blog/2016/02/best-way-to-store-dotfiles-git-bare-repo/

However, I thought it was a little bit hard to keep track of
all my dotfiles that way, without having a defined place to organize
them.

So I decided to change it to a normal repo, associate it to a
dotfiles folder and make use of thoughtbot rcm to handle the symlinks.
Reference: https://thoughtbot.github.io/rcm/